Restructure project layout
