+++ Dockerfile
@@ -3,4 +3,4 @@ WORKDIR /app
 COPY . /app
 RUN pip install --no-cache-dir -r requirements.txt
 EXPOSE 8000
-CMD ["uvicorn", "api_app:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "api.api_app:app", "--host", "0.0.0.0", "--port", "8000"]
