+++ integrations/storage_datalake.py
@@ -1,6 +1,6 @@
 import os
 from typing import List
-from storage import StorageBackend
+from integrations.storage import StorageBackend
 
 
 class DatalakeStorage(StorageBackend):
